click outside slider closes pageslide
body class option required: this change resgisters a click event on the body and slider to determine when the user clicks on the faded area (body pseudo element). Events are regsitered and deregistered per use of the pageslide.

forgot to delete useless code

slider.contains replaces recurrsion

Added event listener to the body only