Add 'psBuffer' property
The psBuffer property is designed to enable users who have additional
styling elements, such as drop shadows, to properly hide the slideout
when not in use. Currently, the slideout only positions itself off the
page by the same number of pixels as it's width. If it is styled to have
a drop shadow, this shadow is still visible, specifying a buffer enables
the user to move it an additional amount off the viewport.